+++ .github/release-notes.md
@@ -1,3 +1,1 @@
-* More fixes for popups/infinite lists
-* Added weapon kill counts tracker
-* Added steamhistory.net to player context menu
+* Added separators in chat/killfeed to show when the map changed